requirements.txt
aiosqlite    ==     0.21.0
alembic       ==    1.14.1
annotated-types ==  0.7.0
anyio          ==   4.8.0
certifi        ==   2025.1.31
distro        ==    1.9.0
greenlet      ==    3.1.1
h11           ==    0.14.0
httpcore      ==    1.0.7
httpx         ==    0.28.1
idna          ==    3.10
jiter         ==    0.8.2
Mako          ==    1.3.9
MarkupSafe    ==    3.0.2
openai        ==    1.61.1
pip           ==    25.0
pyaes         ==    1.6.1
pyasn1        ==    0.6.1
pydantic      ==    2.10.6
pydantic_core ==    2.27.2
python-dotenv ==    1.0.1
rsa           ==    4.9
sniffio       ==    1.3.1
SQLAlchemy    ==    2.0.38
Telethon      ==    1.38.1
tqdm          ==    4.67.1
typing_extensions== 4.12.2
